migrate sbt build definition from Build.scala to build.sbt
